+++ packages/simple-taro-fast-framework/CHANGELOG.md
@@ -3,6 +3,87 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.246](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.245...simple-taro-fast-framework@1.0.246) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+## [1.0.245](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.244...simple-taro-fast-framework@1.0.245) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+## [1.0.244](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.243...simple-taro-fast-framework@1.0.244) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+## [1.0.243](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.242...simple-taro-fast-framework@1.0.243) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+## [1.0.242](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.241...simple-taro-fast-framework@1.0.242) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+## [1.0.241](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.240...simple-taro-fast-framework@1.0.241) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+## [1.0.240](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.239...simple-taro-fast-framework@1.0.240) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+## [1.0.239](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.238...simple-taro-fast-framework@1.0.239) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+## [1.0.238](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.237...simple-taro-fast-framework@1.0.238) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+## [1.0.237](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.236...simple-taro-fast-framework@1.0.237) (2022-07-27)
+
+**Note:** Version bump only for package simple-taro-fast-framework
+
+
+
+
+
+
 ## [1.0.236](https://github.com/kityandhero/taro-fast-framework/compare/simple-taro-fast-framework@1.0.235...simple-taro-fast-framework@1.0.236) (2022-07-26)
 
 **Note:** Version bump only for package simple-taro-fast-framework