+++ packages/taro-fast-framework/CHANGELOG.md
@@ -3,6 +3,88 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.248](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.247...taro-fast-framework@1.0.248) (2022-07-27)
+
+**Note:** Version bump only for package taro-fast-framework
+
+
+
+
+
+## [1.0.247](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.246...taro-fast-framework@1.0.247) (2022-07-27)
+
+**Note:** Version bump only for package taro-fast-framework
+
+
+
+
+
+## [1.0.246](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.245...taro-fast-framework@1.0.246) (2022-07-27)
+
+
+### Performance Improvements
+
+* **simple-taro-fast-framework:** update ([2126bd7](https://github.com/kityandhero/taro-fast-framework/commit/2126bd7b567ddcd7420eb1ef96310e138caff188))
+
+
+
+## [1.0.245](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.244...taro-fast-framework@1.0.245) (2022-07-27)
+
+**Note:** Version bump only for package taro-fast-framework
+
+
+
+
+
+## [1.0.244](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.243...taro-fast-framework@1.0.244) (2022-07-27)
+
+**Note:** Version bump only for package taro-fast-framework
+
+
+
+
+
+## [1.0.243](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.242...taro-fast-framework@1.0.243) (2022-07-27)
+
+**Note:** Version bump only for package taro-fast-framework
+
+
+
+
+
+## [1.0.242](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.241...taro-fast-framework@1.0.242) (2022-07-27)
+
+**Note:** Version bump only for package taro-fast-framework
+
+
+
+
+
+## [1.0.241](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.240...taro-fast-framework@1.0.241) (2022-07-27)
+
+**Note:** Version bump only for package taro-fast-framework
+
+
+
+
+
+## [1.0.240](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.239...taro-fast-framework@1.0.240) (2022-07-27)
+
+**Note:** Version bump only for package taro-fast-framework
+
+
+
+
+
+## [1.0.239](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.238...taro-fast-framework@1.0.239) (2022-07-27)
+
+**Note:** Version bump only for package taro-fast-framework
+
+
+
+
+
+
 ## [1.0.238](https://github.com/kityandhero/taro-fast-framework/compare/taro-fast-framework@1.0.237...taro-fast-framework@1.0.238) (2022-07-26)